requirements.txt

ipython
jupyter
numpy
scipy
matplotlib
flake8
scandir
graphviz
click
python-igraph
networkx
sklearn
cairocffi
